cleanup and small fixes to layout
